requirements.txt
anyio==3.6.1
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
fastapi==0.82.0
gitdb==4.0.9
GitPython==3.1.27
gotrue==0.5.0
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.21.3
idna==3.3
importlib-metadata==4.12.0
invoke==1.7.1
jaraco.classes==3.2.2
keyring==23.9.1
more-itertools==8.14.0
packaging==21.3
pkginfo==1.8.3
postgrest-py==0.10.2
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.9.0
python-semantic-release==7.28.1
PyYAML==6.0
readme-renderer==37.1
realtime==0.0.4
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.19.1
storage3==0.3.4
supabase==0.5.8
tomlkit==0.10.2
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
webencodings==0.5.1
websockets==9.1
zipp==3.8.1
gitpython=3.1.27